Move process to custom hook
